# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Tracks api",
        "description": "Управление сущностями проекта трекера произведений.",
        "contact": {
            "email": "xenx.efy@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost/async",
            "description": "Main server"
        }
    ],
    "paths": {
        "/composers": {
            "get": {
                "tags": [
                    "Composer"
                ],
                "summary": "Получение композиторов.",
                "operationId": "getComposers",
                "responses": {
                    "200": {
                        "description": "Success response"
                    }
                }
            },
            "post": {
                "tags": [
                    "Composer"
                ],
                "summary": "Создает композитора.",
                "operationId": "storeComposer",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "composerName"
                                ],
                                "properties": {
                                    "composerName": {
                                        "description": "Название произведения",
                                        "type": "string",
                                        "example": "Бетховен"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success response"
                    },
                    "201": {
                        "description": "Success creation"
                    }
                }
            }
        },
        "/composers/{id}": {
            "put": {
                "tags": [
                    "Composer"
                ],
                "summary": "Обновляет информацию о композиторе.",
                "operationId": "updateComposer",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id обновляемого композитора",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Формат входящих данных",
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "composerName": {
                                        "description": "Название произведения",
                                        "type": "string",
                                        "example": "Adazio"
                                    },
                                    "composerId": {
                                        "description": "Id композитора произведения",
                                        "type": "integer",
                                        "example": 12
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success response"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Composer"
                ],
                "summary": "Удалить композитора.",
                "operationId": "deleteComposer",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id удаляемого композитора.",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success response"
                    }
                }
            }
        },
...
    },
...
    "security": [
        []
    ]
}